refactor: refactor home gamecards
