+++ src/pages/Home/components/GameCards/index.js
@@ -1,0 +1,1 @@
+export { default as GameCards } from './GameCards';
