+++ src/pages/Home/components/index.js
@@ -1,0 +1,2 @@
+export * from './GameCards';
+export * from './index';
